[SWARM-1646] fix the S2I flow (#22)
Since the migration of tests to Arquillian Cube, the `name-service`
and `greeting-service` Maven modules produce two JARs: the usual
`-swarm.jar` uberjar, and a `-classes.jar` JAR that contains the
classes portion of the WAR. This is needed because the `tests` module
depends on these `-classes.jar` JARs so that it can configure
Fabric8 Maven plugin to create an aggregate OpenShift YAML, and that
is needed so that Arquillian Cube can deploy both services before
the test.

At the end of the S2I build, both these 2 JARs are copied to
`/deployments`. Later, when starting the container, the `run-java.sh`
script looks into this directory and expects to find a single JAR.
If it can't find 1 and precisely 1 JAR, it fails.

There are 2 options to fix it (that I know about):
- add an environment variable `ARTIFACT_COPY_ARGS` to the `BuildConfig`
  so that only 1 JAR is copied to `/deployments` at the end of the build
- add an environment variable `JAVA_APP_JAR` to the pod template
  of the `DeploymentConfig`, so that the `run-java.sh` script knows
  which JAR should be executed

Since it's possible to use wildcards in `ARTIFACT_COPY_ARGS` but not in
`JAVA_APP_JAR`, this commit uses the 1st option.